PIM-2320: Berries Field Management - Add Field - Re-test Add Field for Grain and Forage
