+++ cirras-underwriting-war/cirras-underwriting-war/src/main/angular/src/app/components/inventory/add-field/add-field.component.html
@@ -60,7 +60,7 @@
             [fieldList]="fieldList.collection"
             [insurancePlanId]="dataReceived.insurancePlanId"
             [searchBy]="addFieldForm.controls.choiceSelected.value"
-            [selectedCommodity]="dataReceived.berries.selectedCommodity"
+            [selectedCommodity]="(dataReceived.berries && dataReceived.berries.selectedCommodity) ? dataReceived.berries.selectedCommodity : null"
             (fieldChanged)="onFieldReceived($event)"></field-list>
     </div>
