+++ cirras-underwriting-war/cirras-underwriting-war/src/main/angular/src/app/components/inventory/add-field/add-field.component.scss
