+++ cirras-underwriting-war/cirras-underwriting-war/src/main/angular/src/app/components/inventory/add-field/field-validations/field-validations.component.html
