+++ cirras-underwriting-war/cirras-underwriting-war/src/main/angular/src/app/components/inventory/add-field/legal-land-list/legal-land-list.component.ts
