--- cirras-underwriting-war/cirras-underwriting-war/src/main/angular/src/app/components/inventory/add-land/add-land.component.scss
@@ -1,99 +1,0 @@
-  .grid-container {
-    display: grid;
-    grid-template-columns: auto auto; // auto;
-    padding: 10px 0;
-    //justify-content: start;
-    align-items: center;
-  
-    .grid-item {
-
-      ::ng-deep  .mdc-text-field--filled {
-
-        /* change the background color */
-        .mat-mdc-form-field-focus-overlay {
-            background-color: transparent !important;
-        }
-      }
-    }
-  }
-
-
-.row-group-left {
-    // padding: 10px 0;
-    display: flex;
-    flex-direction: column;
-    align-items: flex-start;
-
-  .legal-land-container {
-    display: grid;
-    grid-auto-rows: minmax( 2em, auto );
-
-    &.legal-land-info {
-      grid-template-columns: 1fr 2fr 2fr 5fr;
-    }
-
-    &.field-info {
-      grid-template-columns: 1fr 2fr 2fr 3fr;
-    }
-
-    &.error-messages {
-      grid-template-columns: 1fr 20fr;
-    }
-
-    div {
-        display: flex;
-        justify-content: flex-start;
-        align-items: center;
-    }
-
-    .center {
-        justify-content: center;
-    }
-
-    .header {
-        font-weight: bold;            
-    }
-
-    .border-right {
-        border-right: 1px solid #ccc;          
-    }
-
-    .border-top {
-        border-top: 1px solid #ccc;          
-    }
-
-    .border-left {
-        border-left: 1px solid #ccc;          
-    }
-
-    .justify-center {
-        justify-content: center;   
-    }
-
-    .border-bottom {
-        border-bottom: 1px solid #ccc;
-    }
-  }
-
-  .cuw-table-wrapper {
-    max-height: 200px;
-    border-top: 1px solid #ccc;
-    border-bottom: 1px solid #ccc;
-    width: 770px;
-    overflow-y: auto;
-  }
-
-}
-
-.add-land-rb-group {
-  &.radio-group {
-    display: flex;
-    flex-direction: row;
-    margin: 15px 0 25px 0;
-    align-items: flex-start;
-  }
-  
-  &.radio-button {
-    margin-right: 50px;
-  }
-}
